requirements.txt
aistudio-sdk==0.3.8
albucore==0.0.13
albumentations==1.4.10
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
bce-python-sdk==0.9.60
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
distro==1.9.0
filelock==3.25.0
fire==0.7.1
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
groq==0.37.1
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ImageIO==2.37.2
imagesize==2.0.0
imgaug==0.4.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.10
langchain-core==1.2.17
langchain-groq==1.1.2
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.13
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.34.0
networkx==3.6.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
paddleocr==2.9.1
paddlepaddle==2.6.2
paddlex==3.4.2
pandas==3.0.1
pillow==12.1.1
prettytable==3.17.0
protobuf==3.20.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdfium2==5.5.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.2
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tomli==2.4.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.1
wcwidth==0.6.0
xxhash==3.6.0
zstandard==0.25.0
